% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_settings.R
\name{add_workflow_info}
\alias{add_workflow_info}
\alias{add_workflow_info}
\alias{bm_settings2pecan_settings}
\title{Add workflow specific info to settings list for benchmarking}
\usage{
add_workflow_info(settings)

bm_settings2pecan_settings(bm.settings)
}
\arguments{
\item{settings}{settings or multisettings object}

\item{settings}{settings or multisettings object}
}
\author{
Betsy Cowdery

Betsy Cowdery
}
