% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.data.R
\name{align.data}
\alias{align.data}
\title{Align timeseries data}
\usage{
align.data(model_full, obvs_full, dat_vars, start_year, end_year,
  align_method = "match.timestep")
}
\arguments{
\item{model_full}{data.frame}

\item{obvs_full}{data.frame}

\item{dat_vars}{data.frame}

\item{start_year}{numeric}

\item{end_year}{numeric}
}
\value{
dat
}
\author{
Betsy Cowdery
}

