% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_by_first_observation.R
\name{align_by_first_observation}
\alias{align_by_first_observation}
\title{align_first_observation}
\usage{
align_by_first_observation(observation_one, observation_two, custom_table)
}
\arguments{
\item{observation_one}{a vector of plant functional types, or species. Provides species/pft names.}

\item{observation_two}{another vector of plant functional types, or species. Provides the order.}

\item{custom_table}{a table that either maps two pft's to one another or maps custom species codes to bety id codes.
In the second case, must be passable to match_species_id.}
}
\value{
\code{vector} Returns a vector of PFT's/species from observation_one that matches the order of observation_two
}
\description{
align_first_observation
}
\examples{

observation_one <- c("AMCA3", "AMCA3", "AMCA3", "AMCA3")
observation_two <- c("a", "b", "a", "a")

table <- list()
table$plant_functional_type_one <- c("AMCA3", "AMCA3", "ARHY", "ARHY")
table$plant_functional_type_two <- c("a", "a", "b", "b") # PFT groupings
table <- as.data.frame(table)

aligned <- align_by_first_observation(
  observation_one = observation_one,
  observation_two = observation_two,
  custom_table = table
)

# aligned should be a vector '[1] "AMCA3" "ARHY"  "AMCA3" "AMCA3"'
}
\author{
Tempest McCabe
}
