% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_by_obeservation_two.R
\name{align_by_observation_two}
\alias{align_by_observation_two}
\title{align_by_observation_two}
\usage{
align_by_observation_two(observation_one, observation_two, custom_table)
}
\arguments{
\item{observation_one}{a vector of plant fucntional types, or species}

\item{observation_two}{anouther vector of plant fucntional types, or species}

\item{custom_table}{a table that either maps two pft's to one anouther or maps custom species codes to bety id codes. 
In the second case, must be passable to match_species_id.}
}
\examples{

observation_one<-c("AMCA3","AMCA3","AMCA3","AMCA3")
observation_two<-c("a", "b", "a", "a")
table<-list()
table$plant_functional_type_one<- c("AMCA3","AMCA3","ARHY", "ARHY")
table$plant_functional_type_two<- c('a','a','b', 'b') # PFT groupings
table$input_code<-c("AMCA3","AMCA3","ARHY", "ARHY") # Species
table<-as.data.frame(table)

format_one<-"species_USDA_symbol"
format_two<-"plant_funtional_type"

aligned<-align_by_observation_one(observation_one = observation_one, observation_two = observation_two, 
custom_table = table)
# aligned should be a vector '[1] "a" "a" "a" "a"'
}
\author{
Tempest McCabe
}
