% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_data.R
\name{align_data}
\alias{align_data}
\title{Align timeseries data}
\usage{
align_data(model.calc, obvs.calc, var, align_method = "match_timestep")
}
\arguments{
\item{model.calc}{data.frame}

\item{obvs.calc}{data.frame}

\item{var}{data.frame}

\item{align_method}{name of function to use for alignment}
}
\value{
dat
}
\description{
Align timeseries data
}
\author{
Betsy Cowdery
}
