\name{as.psychobench}
\alias{as.psychobench}
\title{Coerce benchmark experiment warehouse to preference table}
\usage{
  as.psychobench(x, comparisons = TRUE)
}
\arguments{
  \item{x}{A \code{\link{warehouse}} object}

  \item{comparisons}{Return preference or performance
  table}
}
\value{
  Data.frame with preference or performance table
}
\description{
  Coerce benchmark experiment warehouse to preference table
}

