\name{benchmark-comptime}
\alias{fittime}
\title{Dummy functions to enable fitting and prediction time as performance...}
\usage{fittime(yhat, y)
predicttime(yhat, y)
}
\description{Dummy functions to enable fitting and prediction time as performance
measures.}
\value{\code{fittime}: Time (User and System) used for the model fitting

\code{predicttime}: Time (User and System) used for the prediction


}
\seealso{\code{\link{benchmark}}}
\alias{benchmark-comptime}
\arguments{\item{yhat}{Ignored}
\item{y}{Ignored}
}
\alias{predicttime}

