\name{benchmark-generics}
\alias{benchmark-generics}
\title{Generic functions in benchmark package}
\description{These generic functions are defined in the package
  benchmark.}
\alias{beplot0}
\usage{beplot0(x, ...)
densityplot(x, ...)
}
\arguments{\item{x}{An object}
\item{...}{Additional arguments}
}
\alias{densityplot}

