% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Automatic benchmark analysis using CITEdb}
\usage{
benchmark(assess)
}
\arguments{
\item{assess}{text files with predicted cell-cell interactions}
}
\value{
the precision-recall curves and tables of precision-recall values
}
\description{
Returns the precision-recall curves and tables of precision-recall values of benchmark analysis using CITEdb
}
\examples{
library(ggplot2)
library(dplyr)
library(yardstick)
library(benchmark)

#load example data
data('Default.sum')
#benchmark one algorithm
benchmark(Default.sum[[1]])

#benchmark multiple algorithms at once
input = do.call(rbind,Default.sum)
benchmark(asses=input)
}
