% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_settings.R
\name{bm_settings2pecan_settings}
\alias{bm_settings2pecan_settings}
\title{Move benchmarking settings back in to original pecan settings object}
\usage{
bm_settings2pecan_settings(bm.settings)
}
\arguments{
\item{bm.settings}{settings or multisettings object}
}
\description{
Move benchmarking settings back in to original pecan settings object
}
\author{
Betsy Cowdery
}
