% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmark.R
\name{bm_validate}
\alias{bm_validate}
\title{Validate benchmark input dataframes}
\usage{
bm_validate(model_df, obs_df)
}
\arguments{
\item{model_df}{data.frame with columns: time (POSIXct), value (numeric)}

\item{obs_df}{data.frame with columns: time (POSIXct), value (numeric)}
}
\value{
invisible(TRUE)
}
\description{
Validate benchmark input dataframes
}
