% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc.benchmark}
\alias{calc.benchmark}
\title{Calculate benchmarking statistics}
\usage{
calc.benchmark(settings, con)
}
\arguments{
\item{settings}{object from start.benchmark.runs (for now - we many only need a few variables)}

\item{con}{database connection}
}
\description{
For each benchmark id, calculate metrics and update benchmarks_ensemble_scores
}
\author{
Betsy Cowdery
}

