% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.metrics.R
\name{calc.metrics}
\alias{calc.metrics}
\title{calc.metrics}
\usage{
calc.metrics(model.bm, obvs.bm, var, metrics, start_year, end_year, bm,
  ensemble.id, model_run)
}
\arguments{
\item{model_run}{}
}
\author{
Betsy Cowdery
}

