% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics.R
\name{calc_metrics}
\alias{calc_metrics}
\title{calc_metrics}
\usage{
calc_metrics(model.calc, obvs.calc, var, metrics, start_year, end_year, bm,
  ensemble.id, model_run)
}
\arguments{
\item{model_run}{}
}
\author{
Betsy Cowdery
}

