% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_settings.R
\name{check_BRR}
\alias{check_BRR}
\title{Check whether a run has been registered as a reference run in BETY}
\usage{
check_BRR(inputfile, con)
}
\arguments{
\item{inputfile}{the PEcAn settings file to be used.}

\item{con}{database connection}
}
\author{
Betsy Cowdery
}
