% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_legal_table.R
\name{check_if_legal_table}
\alias{check_if_legal_table}
\title{check_if_legal_table}
\usage{
check_if_legal_table(table, observation_one, observation_two)
}
\arguments{
\item{table}{a table that either maps two pft's to one anouther or maps custom species codes to bety id codes.
In the second case, must be passable to match_species_id.}

\item{observation_one}{a vector of plant functional types, or species}

\item{observation_two}{anouther vector of plant functional types, or species}
}
\value{
\code{boolean}
}
\description{
check_if_legal_table
}
\details{
Checks if custom_table:
1. is formated correctly
2. is complete (has all of the species/pft's in both observations)
3. is condense-able (Could be represented as a hierachry)
}
\author{
Tempest McCabe
}
