% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_list_of_pfts.R
\name{check_if_list_of_pfts}
\alias{check_if_list_of_pfts}
\title{check_if_list_of_pfts}
\usage{
check_if_list_of_pfts(vars)
}
\arguments{
\item{observation_one}{a vector of plant fucntional types, or species}

\item{observation_two}{anouther vector of plant fucntional types, or species}

\item{custom_table}{a table that either maps two pft's to one anouther or maps custom species codes to bety id codes. 
In the second case, must be passable to match_species_id.}
}
\value{
\code{boolean}
}
\description{
check_if_list_of_pfts
}
\details{
Checks if format contains a variable named "plant_functional_type"
}
\author{
Tempest McCabe
}
