% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_settings.R
\name{clean_settings_BRR}
\alias{clean_settings_BRR}
\title{Cleans PEcAn settings file and prepares the settings to be saved in a reference run record in BETY}
\usage{
clean_settings_BRR(inputfile)
}
\arguments{
\item{inputfile}{the PEcAn settings file to be used.}
}
\description{
Cleans PEcAn settings file and prepares the settings to be saved in a reference run record in BETY
}
\author{
Betsy Cowdery
}
