% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_run.R
\name{clone_run}
\alias{clone_run}
\title{Clone Run}
\usage{
clone_run(settings, model, con)
}
\arguments{
\item{settings}{settings or multisettings object}

\item{model}{model name}

\item{con}{database connection}
}
\description{
Produce a "cloned" run from the settings of a completed run
}
\author{
Betsy Cowdery
}
