% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clone_run_inputs}
\alias{clone_run_inputs}
\title{Clone Run Input Conversion Lookup}
\format{data frame, all columns character
\describe{
 \item{old}{The format of the old data}
 \item{new}{The format of the new data}
 \item{convert_type}{The closest format that can be used to convert between old and new}
 \item{convert_format_id}{Format id of convert_type}
}}
\usage{
clone_run_inputs
}
\description{
A lookup table for determining which file format is necessary for converting
inputs from a run into inputs for a cloned run.
For example for met we need the CF standard 
For IC it depends on pool vs cohort - may need special cases
}
\keyword{datasets}
