% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.benchmark.R
\name{create.benchmark}
\alias{create.benchmark}
\title{Create Bety Benchmarking Records}
\usage{
create.benchmark(settings, bety)
}
\arguments{
\item{settings}{settings list}
}
\value{
updated settings list
}
\description{
Creates records for benchmarks_ensembles, benchmarks, benchmarks_benchmarks_reference_runs, benchmarks_metrics
}
\author{
Betsy Cowdery
}

