% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.BRR.R
\name{create.BRR}
\alias{create.BRR}
\title{Create benchmark reference run and ensemble}
\usage{
create.BRR(ens_wf, con)
}
\arguments{
\item{ens_wf}{table made from joining ensemble and workflow tables}

\item{con}{database connection}
}
\description{
For each benchmark id, calculate metrics and update benchmarks_ensemble_scores
}
\author{
Betsy Cowdery
}

