% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efi_long_to_array.R
\name{efi_long_to_array}
\alias{efi_long_to_array}
\title{Reshape EFI Long Format DataFrame to Matrix of Ensemble Members}
\usage{
efi_long_to_array(df, var = NULL, site = NULL)
}
\arguments{
\item{df}{data.frame in EFI long format.}

\item{var}{character string specifying the target variable to filter (optional).}

\item{site}{character string specifying the target site_id to filter (optional).}
}
\value{
A matrix with timesteps as rows and ensemble members as columns, with a `time` attribute (POSIXct vector).
}
\description{
Takes an EFI long format dataset (containing scenario, datetime/time, site_id/site,
parameter/ensemble/member, variable, prediction/value) and reshapes it into a matrix
of predictions with rows corresponding to timesteps and columns corresponding to ensemble members.
}
