% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_wide2long.R
\name{format_wide2long}
\alias{format_wide2long}
\title{format_wide2long}
\usage{
format_wide2long(out, format, vars_used, time.row)
}
\arguments{
\item{out}{wide format data}

\item{format}{as returned by query.format.vars}

\item{vars_used}{}
}
\value{
list of updated values
}
\description{
format_wide2long
}
\author{
Istem Fer
Function to convert wide format to long format
}
