% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_list_standard.R
\name{get_species_list_standard}
\alias{get_species_list_standard}
\title{get_species_list_standard
Checks if custom_table:
1. is formated correctly
2. is complete (has all of the species/pft's in both observations)
3. is condense-able (Could be represented as a hierachry)}
\usage{
get_species_list_standard(vars)
}
\arguments{
\item{observation_one}{a vector of plant fucntional types, or species}

\item{observation_two}{anouther vector of plant fucntional types, or species}

\item{custom_table}{a table that either maps two pft's to one anouther or maps custom species codes to bety id codes. 
In the second case, must be passable to match_species_id.}
}
\value{
\code{character} Returns "usda", "latin_name", "fia" or "custom"
}
\author{
Tempest McCabe
}
