% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.data}
\alias{load.data}
\title{load.data}
\usage{
load.data(data.path, format, start_year = NA, end_year = NA, site = NA,
  vars_used)
}
\arguments{
\item{data.path}{}

\item{format}{}

\item{start_year}{}

\item{end_year}{}

\item{site}{}
}
\description{
load.data
}
\author{
Betsy Cowdery
Generic function to convert input files containing observational data to
a common PEcAn format.

This should be the same as the read.output function:
This function uses MsTMIP variables except that units of (kg m-2 d-1)
are converted to kg ha-1 y-1.

Currently this function converts

Carbon fluxes: GPP, NPP, NEE, TotalResp, AutoResp, HeteroResp,
DOC_flux, Fire_flux, and Stem (Stem is specific to the BioCro model)

Water fluxes: Evaporation (Evap), Transpiration(TVeg),
surface runoff (Qs), subsurface runoff (Qsb), and rainfall (Rainf).
}

