% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.netcdf.R
\name{load.x_netcdf}
\alias{load.x_netcdf}
\title{load.x_netcdf}
\usage{
load.x_netcdf(data.path, format, site, vars = NULL)
}
\arguments{
\item{data.path}{character vector or list}

\item{format}{list}

\item{site}{list}

\item{vars}{character}

\item{start_year}{numeric}

\item{end_year}{numeric}
}
\author{
Istem Fer
}

