% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_intake.R
\name{load_and_map_data}
\alias{load_and_map_data}
\title{Load and standardize arbitrary tabular data using a YAML mapping configuration}
\usage{
load_and_map_data(data.path, mapping.path)
}
\arguments{
\item{data.path}{character, file path to the tabular data (e.g. .csv)}

\item{mapping.path}{character, file path to the YAML mapping configuration}
}
\value{
A standardized data frame with column names mapped to PEcAn standard vocabulary
}
\description{
Load and standardize arbitrary tabular data using a YAML mapping configuration
}
