% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv.R
\name{load_csv}
\alias{load_csv}
\title{load_csv}
\usage{
load_csv(data.path, format, site, vars = NULL)
}
\arguments{
\item{data.path}{character, file path to the .csv file}

\item{format}{list, config list containing 
\itemize{
  \item \code{header},    numeric for number of header rows in file
  \item \code{skip},       numeric for skip argument of utils::read.csv 
  \item \code{na.strings}, character vector for na.strings argument
                          of utils::read.csv
  }}

\item{site}{list, currently ignored}

\item{vars}{column names to return. If NULL, returns all columns}
}
\value{
a data frame with the processed data of the .csv file
}
\description{
load_csv
}
\author{
Betsy Cowdery
}
