% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(data.path, format, start_year = NA, end_year = NA,
  site = NA, vars.used.index = NULL, ...)
}
\arguments{
\item{data.path}{character}

\item{format}{list}

\item{start_year}{numeric}

\item{end_year}{numeric}

\item{site}{list}
}
\description{
load_data
}
\author{
Betsy Cowdery, Istem Fer, Joshua Mantooth
Generic function to convert input files containing observational data to 
a common PEcAn format.
}
