% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rds.R
\name{load_rds}
\alias{load_rds}
\title{load_rds}
\usage{
load_rds(data.path, format, site, vars = NULL)
}
\arguments{
\item{data.path}{character}

\item{format}{list, not used, for compatibility}

\item{site}{not used, for compatibility}

\item{vars}{optional variable names to load. if NULL, returns all variables in file}
}
\description{
load_rds
}
\author{
Istem Fer
}
