% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tab.R
\name{load_tab_separated_values}
\alias{load_tab_separated_values}
\title{load_tab_separated_values}
\usage{
load_tab_separated_values(data.path, format, site = NULL, vars = NULL)
}
\arguments{
\item{data.path}{character}

\item{format}{list}

\item{site}{list}

\item{start_year}{numeric}

\item{end_year}{numeric}
}
\description{
load_tab_separated_values
}
\author{
Betsy Cowdery, Mike Dietze
}
