% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_timestep.R
\name{match_timestep}
\alias{match_timestep}
\title{Match time step}
\usage{
match_timestep(date.coarse, date.fine, data.fine)
}
\arguments{
\item{date.coarse}{numeric}

\item{date.fine}{numeric}

\item{data.fine}{matrix}
}
\description{
Match time step
}
\author{
Istem Fer
}
