% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.over.larger.timestep.R
\name{mean.over.larger.timestep}
\alias{mean.over.larger.timestep}
\title{Calculate benchmarking statistics}
\usage{
\method{mean}{over.larger.timestep}(date.coarse, date.fine, data.fine)
}
\arguments{
\item{date.coarse}{numeric}

\item{date.fine}{numeric}

\item{data.fine}{data.frame}
}
\author{
Betsy Cowdery, Michael Dietze
}

