% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_over_larger_timestep.R
\name{mean_over_larger_timestep}
\alias{mean_over_larger_timestep}
\title{Calculate benchmarking statistics}
\usage{
mean_over_larger_timestep(date.coarse, date.fine, data.fine)
}
\arguments{
\item{date.coarse}{numeric}

\item{date.fine}{numeric}

\item{data.fine}{data.frame}
}
\description{
Calculate benchmarking statistics
}
\author{
Betsy Cowdery, Michael Dietze
}
