% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.lmDiag.plot.R
\name{metric.lmDiag.plot}
\alias{metric.lmDiag.plot}
\title{metric.lmDiag.plot}
\usage{
metric.lmDiag.plot(metric_dat, var, filename = NA, draw.plot = FALSE)
}
\arguments{
\item{metric_dat}{data.frame}
}
\author{
Betsy Cowdery
}

