% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMC.R, R/metric.PPMC.R
\name{metric.PPMC}
\alias{metric.MSE}
\alias{metric.PPMC}
\title{Pearson Product Moment Correlation}
\usage{
metric.MSE(dat, ...)

metric.MSE(dat, ...)
}
\arguments{
\item{dat}{}

\item{dat}{dataframe}
}
\author{
Betsy Cowdery

Betsy Cowdery
}

