% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.residual.plot.R
\name{metric.residual.plot}
\alias{metric.residual.plot}
\title{Residual Plot}
\usage{
metric.residual.plot(dat, var, filename = NA, draw.plot = FALSE)
}
\arguments{
\item{dat}{}
}
\author{
Betsy Cowdery
}

