% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.scatter.R
\name{metric.scatter.plot}
\alias{metric.scatter.plot}
\title{metric.scatter.plot}
\usage{
metric.scatter.plot(metric_dat, var, filename = NA, draw.plot = FALSE)
}
\arguments{
\item{var}{}
}
\author{
Betsy Cowdery
}

