% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.timeseries.plot.R
\name{metric.timeseries.plot}
\alias{metric.timeseries.plot}
\title{metric.timeseries.plot}
\usage{
metric.timeseries.plot(metric_dat, var, filename = NA, draw.plot = FALSE)
}
\arguments{
\item{metric_dat}{dataframe}

\item{var}{character}
}
\author{
Betsy Cowdery
}

