% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_AME.R
\name{metric_AME}
\alias{metric_AME}
\title{Absolute Maximum Error}
\usage{
metric_AME(dat, calc_method = "single_calc", ...)
}
\arguments{
\item{dat}{dataframe}

\item{calc_method}{}
}
\description{
Absolute Maximum Error
}
\author{
Betsy Cowdery
}
