% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_Bias.R
\name{metric_Bias}
\alias{metric_Bias}
\title{Mean Bias Error}
\usage{
metric_Bias(dat, ...)
}
\arguments{
\item{dat}{dataframe with columns `model` and `obvs`}

\item{...}{ignored}
}
\value{
A numeric value representing the mean bias (mean(model - obvs)).
}
\description{
Mean Bias Error (BIAS)
}
\details{
Computes the mean bias error between model predictions and observations.
}
