% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_cor.R
\name{metric_cor}
\alias{metric_cor}
\title{Correlation Coefficient}
\usage{
metric_cor(dat, calc_method = "single_calc", ...)
}
\arguments{
\item{dat}{dataframe}

\item{calc_method}{}
}
\description{
Correlation Coefficient
}
\author{
Mike Dietze
}
