% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{metric_ensemble_calc}
\alias{metric_ensemble_calc}
\title{Methods of calculation for benchmarking metrics depending on ensemble size}
\format{}
\usage{
metric_ensemble_calc
}
\description{
A lookup table with benchmerking metric names and ids along with the 
approaches that should be taken to obtain metric scores depending on whether 
or not the model simulation contains more than one ensemble member or not.
}
\keyword{datasets}
