% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_Frechet.R
\name{metric_Frechet}
\alias{metric_Frechet}
\title{Frechet Distance}
\usage{
metric_Frechet(metric_dat, ...)
}
\arguments{
\item{metric_dat}{dataframe}

\item{...}{ignored}
}
\description{
Frechet Distance
}
\author{
Betsy Cowdery
}
