% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_lmDiag_plot.R
\name{metric_lmDiag_plot}
\alias{metric_lmDiag_plot}
\title{Linear Regression Diagnostic Plot}
\usage{
metric_lmDiag_plot(metric_dat, var, filename = NA, draw.plot = FALSE)
}
\arguments{
\item{metric_dat}{data.frame}

\item{var}{ignored}

\item{filename}{path to save plot, or NA to not save}

\item{draw.plot}{logical: return plot object?}
}
\description{
Linear Regression Diagnostic Plot
}
\author{
Betsy Cowdery
}
