% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_MSE.R
\name{metric_MSE}
\alias{metric_MSE}
\title{Mean Square Error}
\usage{
metric_MSE(dat, ...)
}
\arguments{
\item{dat}{dataframe}

\item{...}{ignored}
}
\description{
Mean Square Error
}
\author{
Betsy Cowdery
}
