% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_PMU.R
\name{metric_PMU}
\alias{metric_PMU}
\title{Predictive Model Uncertainty (PMU)}
\usage{
metric_PMU(dat, ...)
}
\arguments{
\item{dat}{dataframe with columns `obs_se` and `obs_n`}

\item{...}{ignored}
}
\value{
A numeric value representing the pooled measurement uncertainty.
}
\description{
Predictive Model Uncertainty (PMU)
}
\details{
Calculates the pooled measurement uncertainty.
Requires `obs_se` (standard error) and `obs_n` (replicate counts) in the observation data.
}
