% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_PPMC.R
\name{metric_PPMC}
\alias{metric_PPMC}
\title{Pearson Product Moment Correlation}
\usage{
metric_PPMC(metric_dat, calc_method = "single_calc", ...)
}
\arguments{
\item{metric_dat}{dataframe}

\item{calc_method}{}
}
\description{
Pearson Product Moment Correlation
}
\author{
Betsy Cowdery
}
