% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_RMSE.R
\name{metric_RMSE}
\alias{metric_RMSE}
\title{Root Mean Square Error}
\usage{
metric_RMSE(dat, calc_method = "single_calc", ...)
}
\arguments{
\item{dat}{dataframe}

\item{calc_method}{}
}
\description{
Root Mean Square Error
}
\author{
Betsy Cowdery
}
