% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_scatter_plot.R
\name{metric_scatter_plot}
\alias{metric_scatter_plot}
\title{Scatter Plot}
\usage{
metric_scatter_plot(
  metric_dat,
  var,
  unit = NULL,
  filename = NA,
  draw.plot = is.na(filename)
)
}
\arguments{
\item{metric_dat}{dataframe to plot, with at least columns `model` and `obvs`}

\item{var}{title for the plot}

\item{unit}{measurement unit for the variable, added to the axis labels; NULL to omit}

\item{filename}{path to save plot, or NA to not save}

\item{draw.plot}{logical: Return the plot object?}
}
\description{
Scatter Plot
}
\author{
Betsy Cowdery
}
