% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_timeseries_plot.R
\name{metric_timeseries_plot}
\alias{metric_timeseries_plot}
\title{Timeseries Plot}
\usage{
metric_timeseries_plot(
  metric_dat,
  var,
  unit = NULL,
  filename = NA,
  draw.plot = is.na(filename)
)
}
\arguments{
\item{metric_dat}{dataframe to plot, with at least columns `time`, `model`, `obvs`}

\item{var}{variable name, used as plot title}

\item{filename}{path to save plot, or NA to not save}

\item{draw.plot}{logical: Return the plot object?}
}
\description{
Timeseries Plot
}
\author{
Betsy Cowdery
}
