% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_VariantAnnot:GoogleGenomics.R
\name{micro_getVariantCalls}
\alias{micro_getVariantCalls}
\title{Result :
Unit: seconds
 expr       min       lq        mean      median    uq        max       neval
 myread   1.965495 1.987981  2.345124  2.079625  2.822946  2.869573       5}
\usage{
micro_getVariantCalls()
}
\description{
Extracting the genotypes for interval ranges between - 16051400,16051500 on chromosome 22 using GoogleGenomics approach.
Looking at the time taken to perform this using the GoogleGenomics approach.
}

