% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks_localreadVCF.R
\name{micro_localreadVcf}
\alias{micro_localreadVcf}
\title{benchmark readVcf for select genomic interval, and return the vcf object with timings in a list for a local file.}
\usage{
micro_localreadVcf(chr = "22", start = 16051400, end = 16051500,
  times = 5)
}
\arguments{
\item{chr}{character encoding chromosome in the vcf}

\item{start}{starting location for interval to search for variants}

\item{end}{end location of interval}

\item{times}{passed to \code{\link[microbenchmark]{microbenchmark}}}
}
\value{
a list with components \code{bench} giving timings, \code{data} giving variants, and \code{call} describing call
}
\author{
Shweta Gopaulakrishnan <reshg@channing.harvard.edu>
}

