\name{monks3raw}
\docType{data}
\alias{monks3raw}
\title{monks3 benchmark experiment results}
\description{
Misclassification error and computation time on the data set
\code{monks3} of the candidate algorithms \code{lda}, \code{rf},
\code{knn}, \code{rpart}, \code{svm} and \code{nnet}.

Used in \code{demo("benchplot")}.
}
\usage{monks3raw}
\format{An array of dimension 250 x 6 x 2 x 1 (sampling x algorithms x
  performances x data sets).}
\references{
Manuel J. A. Eugster, Torsten Hothorn and Friedrich Leisch. "Exploratory
and Inferential Analysis of Benchmark Experiments", Technical Report 30,
Department of Statistics, Ludwig-Maximilians-Universitaet Muenchen,
Germany. 2008. http://epub.ub.uni-muenchen.de/4134/
}
\keyword{datasets}
