% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmark.R
\docType{data}
\name{pecan_metric_registry}
\alias{pecan_metric_registry}
\title{Metric Registry for PEcAn.benchmark}
\format{
An object of class \code{environment} of length 7.
}
\usage{
pecan_metric_registry
}
\description{
Metric Registry for PEcAn.benchmark
}
\keyword{datasets}
