% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmark.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot model vs observations time series}
\usage{
plot_time_series(aligned)
}
\arguments{
\item{aligned}{data.frame with columns: time, model, obs}
}
\value{
ggplot object
}
\description{
Plot model vs observations time series
}
