% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.settings.RR.R
\name{read.settings.RR}
\alias{read.settings.RR}
\title{Read settings from database using reference run id}
\usage{
read.settings.RR(settings)
}
\arguments{
\item{settings}{settings or multisettings object}
}
\description{
For each benchmark entry in a (multi)settings object, get run settings using reference run id 
and add to the settings object
}
\author{
Betsy Cowdery
}

