% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_settings.R
\name{read_settings_BRR}
\alias{read_settings_BRR}
\title{Read settings from database using reference run id}
\usage{
read_settings_BRR(settings)
}
\arguments{
\item{settings}{settings or multisettings object}
}
\description{
For each benchmark entry in a (multi)settings object, get run settings using reference run id 
and add to the settings object
}
\author{
Betsy Cowdery
}
