% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmark.R
\name{register_metric}
\alias{register_metric}
\title{Register a new benchmark metric}
\usage{
register_metric(name, func)
}
\arguments{
\item{name}{Character name of the metric}

\item{func}{Function that takes an aligned dataframe and returns a numeric value}
}
\description{
Register a new benchmark metric
}
