% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.bm.ensemble.R
\name{start.bm.ensemble}
\alias{start.bm.ensemble}
\title{Start benchmark ensemble}
\usage{
\method{start}{bm.ensemble}(BRR, con)
}
\arguments{
\item{con}{database connection}
}
\description{
Start Benchmark Ensemble
}
\author{
Betsy Cowdery
}

