% Generated by roxygen2 (4.0.2): do not edit by hand
\name{start.bm.run}
\alias{start.bm.run}
\title{Start benchmark run}
\usage{
\method{start}{bm.run}(BRR.id, database, machine.id)
}
\arguments{
\item{machine.id}{}

\item{benchmark_reference_run.id,}{}
}
\description{
Start Benchmark Run
}
\author{
Betsy Cowdery
}

