\docType{data}
\name{StatlogCharacteristics}
\alias{StatlogCharacteristics}
\title{StatLog dataset characteristics}
\format{proto object 
 $ requirements:function (., ...)   
 $ name        : chr "Statlog" 
 $ reduce      :function (.)   
 $ map         :function (.)   
 parent: proto object}
\description{
  Implementation of the StatLog project dataset
  characteristics.
}
\references{
  See \emph{Eugster et al. (2010)} in
  \code{citation("benchmark")}.

  R. D. King, C. Feng and A. Sutherland. STATLOG:
  Comparison of classification algorithms on large
  real-world problems. Applied Artifical Intelligence, 9,
  1995.
}
\seealso{
  Other dataset.characterization: \code{\link{as.dataset}},
  \code{\link{characterize}},
  \code{\link{plot.DatasetCharacterization}}
}
\keyword{datasets}

