\name{subset.AlgorithmPerformance}
\alias{subset.AlgorithmPerformance}
\title{Return subsets of AlgorithmPerformance objects...}
\usage{\method{subset}{AlgorithmPerformance}(x, datasets, algorithms, performances, samples, ...)
}
\description{Return subsets of \code{AlgorithmPerformance} objects}
\value{An \code{\link{AlgorithmPerformance}} object with just the
selected observations}
\arguments{\item{x}{An \code{\link{AlgorithmPerformance}} object}
\item{datasets}{Selected datasets}
\item{algorithms}{Selected algorithms}
\item{performances}{Selected performances}
\item{samples}{Selected samples}
\item{...}{Ignored}
}

