\name{testres-visualization}
\alias{plot.TestResult}
\title{Visualization methods for (sequential) test results.}
\usage{\method{plot}{TestResult}(x, ...)
}
\description{Visualization methods for (sequential) test results.}
\value{A \code{\link[ggplot2]{ggplot}} object.}
\arguments{\item{x}{An \code{\link{TestResult}} object}
\item{...}{Ignored.}
}

