% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_settings.R
\name{add_workflow_info}
\alias{add_workflow_info}
\title{Add workflow specific info to settings list for benchmarking}
\usage{
add_workflow_info(settings, bety)
}
\arguments{
\item{settings}{settings or multisettings object}

\item{bety}{connection to the database}
}
\description{
Add workflow specific info to settings list for benchmarking
}
\author{
Betsy Cowdery
}
