\name{as.warehouse}
\alias{as.warehouse.mlr.bench.result}
\title{as.warehouse}
\usage{as.warehouse.mlr.bench.result(x, ...)
}
\description{Methods to coerce objects to a benchmark experiment warehouse.}
\details{Coerces a \code{bench.result} object from package \code{mlr} to a
\code{\link{warehouse}} object.}
\value{A \code{\link{warehouse}} object}
\alias{as.warehouse}
\arguments{\item{x}{A \code{bench.result} object from package \code{mlr}}
\item{...}{Ignored}
}

