\name{fittime}
\alias{benchmark-comptime}
\alias{fittime}
\alias{predicttime}
\title{Performance measures}
\usage{
  fittime(yhat, y)

  predicttime(yhat, y)
}
\arguments{
  \item{yhat}{Ignored}

  \item{y}{Ignored}
}
\value{
  Time (User and System) used for the model fitting or
  prediction
}
\description{
  Dummy functions to enable fitting and prediction time as
  performance measures.
}
\seealso{
  \code{\link{benchmark}}
}

