% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.benchmark.R
\name{calc.benchmark}
\alias{calc.benchmark}
\title{Calculate benchmarking statistics}
\usage{
calc.benchmark(bm.ensemble, con)
}
\arguments{
\item{bm.ensemble}{object, either from create.BRR or start.bm.ensemle}

\item{con}{database connection}
}
\description{
For each benchmark id, calculate metrics and update benchmarks_ensemble_scores
}
\author{
Betsy Cowdery
}

