% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.data.R, R/calc.benchmark.R
\name{calc.benchmark}
\alias{align.data}
\alias{calc.benchmark}
\title{Calculate benchmarking statistics}
\usage{
align.data(model, obvs)

calc.benchmark(settings, con)
}
\arguments{
\item{model}{list}

\item{obvs}{list}

\item{settings}{object from start.benchmark.runs (for now - we many only need a few variables)}

\item{con}{database connection}
}
\description{
For each benchmark id, calculate metrics and update benchmarks_ensemble_scores

For each benchmark id, calculate metrics and update benchmarks_ensemble_scores
}
\author{
Betsy Cowdery

Betsy Cowdery
}

