% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc.metrics}
\alias{calc.metrics}
\title{calc.metrics}
\usage{
calc.metrics(data.path, format, model_run, metrics, start_year = NA,
  end_year = NA, site = NA)
}
\arguments{
\item{data.path}{}

\item{format}{}

\item{model_run}{}

\item{metrics}{}

\item{start_year}{}

\item{end_year}{}
}
\description{
calc.metrics
}
\author{
Betsy Cowdery
}

