% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.metrics.R
\name{calc.metrics}
\alias{calc.metrics}
\title{calc.metrics}
\usage{
calc.metrics(data.path, format, vars_used, model_run, metrics, start_year,
  end_year, site)
}
\arguments{
\item{end_year}{}
}
\author{
Betsy Cowdery
}

