% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.metrics.R
\name{calc.metrics}
\alias{calc.metrics}
\title{calc.metrics}
\usage{
calc.metrics(data.path, format, model_run, metrics, start_year = NA,
  end_year = NA, site = NA)
}
\arguments{
\item{end_year}{}
}
\author{
Betsy Cowdery
}

