% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_settings.R
\name{check_BRR}
\alias{check_BRR}
\title{Check whether a run has been registered as a reference run in BETY}
\usage{
check_BRR(settings_xml, con)
}
\arguments{
\item{settings_xml}{cleaned settings to be compared with BRR in the database}

\item{con}{database connection}
}
\description{
Check whether a run has been registered as a reference run in BETY
}
\author{
Betsy Cowdery
}
