% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_list_of_pfts.R
\name{check_if_list_of_pfts}
\alias{check_if_list_of_pfts}
\title{check_if_list_of_pfts}
\usage{
check_if_list_of_pfts(vars)
}
\arguments{
\item{vars}{names to check}
}
\value{
\code{boolean}
}
\description{
Checks if format contains a variable named "plant_functional_type"
}
\author{
Tempest McCabe
}
