% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmark.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute benchmark metrics}
\usage{
compute_metrics(aligned, metrics = c("RMSE", "MAE"))
}
\arguments{
\item{aligned}{data.frame with columns: time, model, obs}

\item{metrics}{character vector of metric names}
}
\value{
data.frame with columns: metric, value
}
\description{
Compute benchmark metrics
}
