% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_benchmark.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute benchmark metrics}
\usage{
compute_metrics(aligned, metrics = c("RMSE", "MAE", "R2"))
}
\arguments{
\item{aligned}{data.frame with columns: model, obvs, time}

\item{metrics}{character vector of metric names}
}
\value{
data.frame in wide format with columns `Site` and each requested metric column per site.
}
\description{
Compute benchmark metrics
}
