% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.BRR.R
\name{create.BRR}
\alias{create.BRR}
\title{Create benchmark reference run and ensemble}
\usage{
create.BRR(ensemble.id, workflow, con)
}
\arguments{
\item{ensemble.id}{id of ensemble run that will be used for benchmarking}

\item{con}{database connection}
}
\description{
For each benchmark id, calculate metrics and update benchmarks_ensemble_scores
}
\author{
Betsy Cowdery
}

