% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_benchmark.R
\name{define_benchmark}
\alias{define_benchmark}
\title{Benchmark Definition: Retrieve or Create Bety Benchmarking Records}
\usage{
define_benchmark(settings, bety)
}
\arguments{
\item{settings}{settings list}

\item{bety}{database connection}
}
\value{
updated settings list
}
\description{
Creates records for benchmarks, benchmarks_benchmarks_reference_runs, benchmarks_metrics
}
\author{
Betsy Cowdery
}
