% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_list_standard.R
\name{get_species_list_standard}
\alias{get_species_list_standard}
\title{get_species_list_standard}
\usage{
get_species_list_standard(vars)
}
\arguments{
\item{observation_one}{a vector of plant fucntional types, or species}

\item{observation_two}{anouther vector of plant fucntional types, or species}

\item{custom_table}{a table that either maps two pft's to one another or maps custom species codes to bety id codes. 
In the second case, must be passable to match_species_id.}
}
\value{
\code{character} Returns "usda", "latin_name", "fia" or "custom"
}
\description{
get_species_list_standard
}
\details{
Returns the format type for convience of use with match_species_id
}
\author{
Tempest McCabe
}
