% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_list_standard.R
\name{get_species_list_standard}
\alias{get_species_list_standard}
\title{get_species_list_standard}
\usage{
get_species_list_standard(vars)
}
\arguments{
\item{vars}{format to be matched}
}
\value{
\code{character} Returns "usda", "latin_name", "fia" or "custom"
}
\description{
Returns the format type for convience of use with match_species_id
}
\author{
Tempest McCabe
}
