% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.csv.R
\name{load.csv}
\alias{load.csv}
\title{load.csv}
\usage{
load.csv(data.path, format, site, vars = NULL)
}
\arguments{
\item{data.path}{character}

\item{format}{list}

\item{site}{list}

\item{start_year}{numeric}

\item{end_year}{numeric}
}
\author{
Betsy Cowdery
}

