% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.csv}
\alias{load.csv}
\title{load.csv}
\usage{
load.csv(data.path, format, start_year, end_year, site)
}
\arguments{
\item{data.path}{}

\item{format}{}

\item{start_year}{}

\item{end_year}{}

\item{site}{}
}
\description{
load.csv
}
\author{
Betsy Cowdery
}

