% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{load.data}
\alias{load.data}
\title{load.data}
\usage{
load.data(data.path, format, start_year = NA, end_year = NA, site = NA,
  vars.used.index, time.row = NULL)
}
\arguments{
\item{data.path}{character}

\item{format}{list}

\item{start_year}{numeric}

\item{end_year}{numeric}

\item{site}{list}
}
\author{
Betsy Cowdery
Generic function to convert input files containing observational data to 
a common PEcAn format.
}

