% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.csv.R, R/load.data.R
\name{load.data}
\alias{load.csv}
\alias{load.data}
\title{load.data}
\usage{
load.csv(data.path, format, start_year, end_year, site)

load.data(data.path, format, start_year = NA, end_year = NA, site = NA)
}
\arguments{
\item{site}{}

\item{site}{}
}
\description{
Generic function to convert input files containing observational data to 
a common PEcAn format.
}
\details{
This should be the same as the read.output function:
This function uses MsTMIP variables except that units of (kg m-2 d-1)  
are converted to kg ha-1 y-1.

Currently this function converts

Carbon fluxes: GPP, NPP, NEE, TotalResp, AutoResp, HeteroResp,
DOC_flux, Fire_flux, and Stem (Stem is specific to the BioCro model)

Water fluxes: Evaporation (Evap), Transpiration(TVeg),
surface runoff (Qs), subsurface runoff (Qsb), and rainfall (Rainf).
}
\author{
Betsy Cowdery

Betsy Cowdery
}

