% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv.R
\name{load_csv}
\alias{load_csv}
\title{load_csv}
\usage{
load_csv(data.path, format, site, vars = NULL)
}
\arguments{
\item{data.path}{character}

\item{format}{list}

\item{site}{list}

\item{vars}{column names to return. If NULL, returns all columns}
}
\description{
load_csv
}
\author{
Betsy Cowdery
}
