% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load data}
\usage{
load_data(
  data.path,
  format,
  start_year = NA,
  end_year = NA,
  site = NA,
  vars.used.index = NULL,
  ...
)
}
\arguments{
\item{data.path}{character}

\item{format}{list}

\item{start_year}{numeric}

\item{end_year}{numeric}

\item{site}{list}

\item{vars.used.index}{which variables to use? If NULL, these are taken from `format`}

\item{...}{further arguments, currently ignored}
}
\description{
Generic function to convert input files containing observational data to
a common PEcAn format.
}
\author{
Betsy Cowdery, Istem Fer, Joshua Mantooth
}
