% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tab.R
\name{load_tab_separated_values}
\alias{load_tab_separated_values}
\title{Load files with mime-type 'text/tab-separated-values'}
\usage{
load_tab_separated_values(data.path, format, site = NULL, vars = NULL)
}
\arguments{
\item{data.path}{character}

\item{format}{list}

\item{site}{list}

\item{vars}{variable names to load. If NULL, loads all columns}
}
\description{
Load files with mime-type 'text/tab-separated-values'
}
\author{
Betsy Cowdery, Mike Dietze
}
