% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_netcdf.R
\name{load_x_netcdf}
\alias{load_x_netcdf}
\title{Load from netCDF}
\usage{
load_x_netcdf(data.path, format, site, vars = NULL)
}
\arguments{
\item{data.path}{character vector or list of paths to NetCDF files}

\item{format}{list containing format info (e.g. na.strings)}

\item{site}{list site info (optional)}

\item{vars}{character vector of variables to load}
}
\value{
A data.frame with the requested variables and a `time` column (POSIXct)
}
\description{
Load from netCDF
}
\author{
Istem Fer, Anshul Jain
}
