% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_netcdf.R
\name{load_x_netcdf}
\alias{load_x_netcdf}
\title{Load from netCDF}
\usage{
load_x_netcdf(data.path, format, site, vars = NULL)
}
\arguments{
\item{data.path}{character vector or list}

\item{format}{list}

\item{site}{list}

\item{vars}{character}
}
\description{
Load from netCDF
}
\author{
Istem Fer
}
