% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.timestep.R
\name{match.timestep}
\alias{match.timestep}
\title{Match time step}
\usage{
match.timestep(date.coarse, date.fine, data.fine)
}
\arguments{
\item{date.coarse}{numeric}

\item{date.fine}{numeric}

\item{data.fine}{matrix}
}
\author{
Istem Fer
}

