% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_Coverage.R
\name{metric_Coverage}
\alias{metric_Coverage}
\title{Prediction Interval Coverage}
\usage{
metric_Coverage(dat, ...)
}
\arguments{
\item{dat}{dataframe with columns `model_q025` and `model_q975`}

\item{...}{ignored}
}
\value{
A numeric value representing the fraction of observations that fall within the 95% prediction interval.
}
\description{
Prediction Interval Coverage
}
\details{
Measures the fraction of observations that fall within the model's
stated 95% prediction interval.
}
