% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_CRPS.R
\name{metric_CRPS}
\alias{metric_CRPS}
\title{Continuous Rank Probability Score}
\usage{
metric_CRPS(metric_dat, calc_method = "ensemble_calc", ...)
}
\arguments{
\item{metric_dat}{dataframe}

\item{calc_method}{}
}
\description{
Continuous Rank Probability Score
}
\author{
Betsy Cowdery
}
