% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_MAE.R
\name{metric_MAE}
\alias{metric_MAE}
\title{Mean Absolute Error}
\usage{
metric_MAE(dat, ...)
}
\arguments{
\item{dat}{dataframe}

\item{...}{ignored}
}
\description{
Mean Absolute Error
}
\author{
Betsy Cowdery
}
