% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_R2.R
\name{metric_R2}
\alias{metric_R2}
\title{Coefficient of Determination (R2)}
\usage{
metric_R2(metric_dat, ...)
}
\arguments{
\item{metric_dat}{dataframe}

\item{...}{ignored}
}
\description{
Coefficient of Determination (R2)
}
\author{
Betsy Cowdery
}
