% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_RAE.R
\name{metric_RAE}
\alias{metric_RAE}
\title{Relative Absolute Error}
\usage{
metric_RAE(metric_dat, ...)
}
\arguments{
\item{metric_dat}{dataframe}

\item{...}{ignored}
}
\description{
Relative Absolute Error
}
\author{
Betsy Cowdery
}
