% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_residual_plot.R
\name{metric_residual_plot}
\alias{metric_residual_plot}
\title{Residual Plot}
\usage{
metric_residual_plot(dat, var, filename = NA, draw.plot = FALSE)
}
\arguments{
\item{dat}{}
}
\author{
Betsy Cowdery
}

