% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_residual_plot.R
\name{metric_residual_plot}
\alias{metric_residual_plot}
\title{Residual Plot}
\usage{
metric_residual_plot(metric_dat, var, filename = NA,
  draw.plot = is.na(filename))
}
\arguments{
\item{draw.plot}{}
}
\author{
Betsy Cowdery
}
