% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_scatter_plot.R
\name{metric_scatter_plot}
\alias{metric_scatter_plot}
\title{Scatter Plot}
\usage{
metric_scatter_plot(metric_dat, var, filename = NA, draw.plot = FALSE)
}
\arguments{
\item{var}{}
}
\author{
Betsy Cowdery
}

