% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_timeseries_plot.R
\name{metric_timeseries_plot}
\alias{metric_timeseries_plot}
\title{Timeseries Plot}
\usage{
metric_timeseries_plot(metric_dat, var, filename = NA,
  draw.plot = is.na(filename))
}
\description{
Timeseries Plot
}
\author{
Betsy Cowdery
}
