% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{micro_getVariantCalls}
\alias{micro_getVariantCalls}
\title{benchmark GG getVariantCalls for select genomic interval, and return the vcf object with timings in a list}
\usage{
micro_getVariantCalls(datasetId = "10473108253681171589", chr = "22",
  start = 16051400, end = 16051500, times = 5, fields = NULL,
  converter = c)
}
\description{
benchmark GG getVariantCalls for select genomic interval, and return the vcf object with timings in a list
}

