% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratv3.R
\name{sce_to_seuratv3}
\alias{sce_to_seuratv3}
\title{SeuratV3 convertion}
\usage{
sce_to_seuratv3(obj)
}
\arguments{
\item{obj}{SingleCellExperiment objcet to convert into a Seurat V3}
}
\description{
SeuratV3 convertion
}
