\name{plot.TestResult}
\alias{plot.TestResult}
\title{Visualization methods for (sequential) test results.}
\usage{
  \method{plot}{TestResult} (x, ...)
}
\arguments{
  \item{x}{An \code{\link{TestResult}} object}

  \item{...}{Ignored.}
}
\value{
  A \code{\link[ggplot2]{ggplot}} object.
}
\description{
  Visualization methods for (sequential) test results.
}

